tower_allowed_hosts

Module matcher

module for matcher

Structs

And

And matcher which matches only when both left and right matches

Any

Any matcher which always returns true and matches any host

Or

Or matcher which matches when either left and right matches

Traits

KeyValueMatcher

Trait for matching the presence and values of parameters in a Forwarded header.

Matcher

Trait for matching against the value