tower_allowed_hosts::matcher

Trait Matcher

pub trait Matcher {
    // Required method
    fn matches_value(&self, value: &str) -> bool;
}

Trait for matching against the value

A Matcher is responsible for checking whether a given value is consider to match with a provided matcher

Required Methods

fn matches_value(&self, value: &str) -> bool

Checks if provided value matches according to matcher