tower_allowed_hosts

Module service

module for layer, service and future

Structs

AllowedHost

Allowed hosts service that wraps the inner service and validates the request host.

AllowedHostFuture

Future for AllowedHost service.

AllowedHostLayer

A layer that validates and allows incoming requests based on their host.