Crate tower_allowed_hosts

Crate which provides allowed hosts layer for tower based service

To use crate with any service you can create layer with any matcher

Examples

let layer = tower_allowed_hosts::AllowedHostLayer::new("example.com");

Check README.MD or documentation of AllowedHostLayer for more detailed information

Modules

error

module for error

matcher

module for matcher

service

module for layer, service and future

Structs

AllowedHostLayer

A layer that validates and allows incoming requests based on their host.

Host

Struct which holds value of host

Enums

Error

Enum for different error