tower_allowed_hosts::matcher

Trait KeyValueMatcher

pub trait KeyValueMatcher {
    // Required method
    fn matches_key_value(&self, values: &HashMap<String, String>) -> bool;
}

Trait for matching the presence and values of parameters in a Forwarded header.

When AllowedHostLayer is configured with a KeyValueMatcher, it will only consider the host= value from a Forwarded header if the matcher determines that the header’s parameters are acceptable.

The matcher receives a map of all key–value pairs in the Forwarded entry (e.g. for=...;by=...;host=...;token=value)

Required Methods

fn matches_key_value(&self, values: &HashMap<String, String>) -> bool

Checks if provided value matches according to matcher