tower_allowed_hosts::matcher
Trait Matcher

pub trait Matcher {
    // Required method
    fn matches_value(&self, value: &str) -> bool;
}

Trait for checking if a given value matches a specific pattern.

Matcher trait is currently used in two places checking Host header as well as Forwarded header value. Both can have custom Matcher which differs from each others

The Matcher trait is designed to be highly versatile and extensible, allowing it to be implemented for a variety of pattern matching scenarios. Beyond simple string comparisons Matcher can be extended to support complex operations according to requirements

Required Methods

fn matches_value(&self, value: &str) -> bool

Checks if provided value matches according to matcher